+++ requirements.txt
@@ -13,6 +13,7 @@ jupyter-server==1.23.4
 jupyter==1.0.0
 jupyterlab-widgets==3.0.4
 keras==2.11.0
+langdetect==1.0.9
 matplotlib-inline==0.1.6
 matplotlib==3.6.3
 mypy-extensions==0.4.3
@@ -38,7 +39,6 @@ requests==2.31.0
 scikit-learn==1.2.0
 scipy==1.9.3
 seaborn==0.12.2
-spacy-language-detection==0.2.1
 spacy==3.5.3
 SQLAlchemy==2.0.9
 SQLAlchemy[mypy]==2.0.9
